requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.27.0
supabase==2.3.7
python-dotenv==1.0.1
google-generativeai==0.4.1
requests==2.31.0
